requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
alembic==1.7.7
asgiref==3.5.0
astroid==2.11.7
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.26.0
atomicwrites==1.4.1
attrs==22.1.0
Babel==2.10.3
bcrypt==3.2.2
black==22.6.0
Brotli==1.0.9
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
classify-imports==4.2.0
cleo==0.8.1
click==8.0.4
clikit==0.6.2
colorama==0.4.5
ConfigArgParse==1.5.3
coverage==6.4.3
crashtest==0.3.1
cryptography==36.0.2
databases==0.6.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
Faker==14.0.0
fastapi==0.75.0
fastapi-admin==1.0.4
filelock==3.6.0
flake8==5.0.4
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httpcore==0.13.2
httptools==0.4.0
httpx==0.18.1
identify==2.5.3
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
iso8601==1.0.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==23.8.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
locust==2.11.0
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
mock-alchemy==0.2.5
msgpack==1.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0
passlib==1.7.4
pastel==0.2.1
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pipenv==2022.8.14
pkginfo==1.8.3
platformdirs==2.5.1
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
pre-commit==2.20.0
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pylev==1.4.0
pylint==2.14.5
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.2
pytest-async-sqlalchemy==0.2.0
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.2
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
redis==4.3.4
reorder-python-imports==3.8.2
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
roundrobin==0.0.4
sentry-sdk==1.9.0
shellingham==1.5.0
six==1.16.0
SQLAlchemy==1.4.40
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.3
tortoise-orm==0.19.2
ujson==5.4.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.13.4
virtualenv-clone==0.5.7
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
